// backend/api/src/schemas/learning.d.ts
export declare const createCourseSchema: z.ZodObject<{
    code: z.ZodOptional<z.ZodString>;
    title: z.ZodString;
    title_en: z.ZodOptional<z.ZodString>;
    description: z.ZodOptional<z.ZodString>;
    description_en: z.ZodOptional<z.ZodString>;
    category: z.ZodOptional<z.ZodString>;
    duration_hours: z.ZodOptional<z.ZodNumber>;
    skill_level: z.ZodOptional<z.ZodEnum<["beginner", "intermediate", "advanced", "expert"]>>;
    provider: z.ZodOptional<z.ZodString>;
    is_mandatory: z.ZodDefault<z.ZodBoolean>;
    is_active: z.ZodDefault<z.ZodBoolean>;
    is_certification: z.ZodDefault<z.ZodBoolean>;
    external_url: z.ZodUnion<[z.ZodOptional<z.ZodString>, z.ZodLiteral<"">]>;
    thumbnail_url: z.ZodUnion<[z.ZodOptional<z.ZodString>, z.ZodLiteral<"">]>;
    language: z.ZodDefault<z.ZodString>;
    tags: z.ZodOptional<z.ZodArray<z.ZodString, "many">>;
    prerequisites: z.ZodOptional<z.ZodArray<z.ZodString, "many">>;
    created_by: z.ZodOptional<z.ZodString>;
}, "strip", z.ZodTypeAny, {
    title: string;
    is_active: boolean;
    language: string;
    is_mandatory: boolean;
    is_certification: boolean;
    code?: string | undefined;
    description?: string | undefined;
    category?: string | undefined;
    created_by?: string | undefined;
    description_en?: string | undefined;
    provider?: string | undefined;
    tags?: string[] | undefined;
    duration_hours?: number | undefined;
    skill_level?: "beginner" | "intermediate" | "advanced" | "expert" | undefined;
    title_en?: string | undefined;
    external_url?: string | undefined;
    thumbnail_url?: string | undefined;
    prerequisites?: string[] | undefined;
}, {
    title: string;
    code?: string | undefined;
    description?: string | undefined;
    is_active?: boolean | undefined;
    category?: string | undefined;
    created_by?: string | undefined;
    description_en?: string | undefined;
    provider?: string | undefined;
    language?: string | undefined;
    is_mandatory?: boolean | undefined;
    tags?: string[] | undefined;
    duration_hours?: number | undefined;
    skill_level?: "beginner" | "intermediate" | "advanced" | "expert" | undefined;
    title_en?: string | undefined;
    is_certification?: boolean | undefined;
    external_url?: string | undefined;
    thumbnail_url?: string | undefined;
    prerequisites?: string[] | undefined;
}>;
export declare const updateCourseSchema: z.ZodObject<{
    title: z.ZodOptional<z.ZodString>;
    title_en: z.ZodOptional<z.ZodString>;
    description: z.ZodOptional<z.ZodString>;
    description_en: z.ZodOptional<z.ZodString>;
    category: z.ZodOptional<z.ZodString>;
    duration_hours: z.ZodOptional<z.ZodNumber>;
    skill_level: z.ZodOptional<z.ZodEnum<["beginner", "intermediate", "advanced", "expert"]>>;
    provider: z.ZodOptional<z.ZodString>;
    is_mandatory: z.ZodOptional<z.ZodBoolean>;
    is_active: z.ZodOptional<z.ZodBoolean>;
    is_certification: z.ZodOptional<z.ZodBoolean>;
    external_url: z.ZodUnion<[z.ZodOptional<z.ZodString>, z.ZodLiteral<"">]>;
    thumbnail_url: z.ZodUnion<[z.ZodOptional<z.ZodString>, z.ZodLiteral<"">]>;
    language: z.ZodOptional<z.ZodString>;
    tags: z.ZodOptional<z.ZodArray<z.ZodString, "many">>;
    prerequisites: z.ZodOptional<z.ZodArray<z.ZodString, "many">>;
    status: z.ZodOptional<z.ZodEnum<["draft", "published", "archived"]>>;
}, "strip", z.ZodTypeAny, {
    description?: string | undefined;
    status?: "draft" | "published" | "archived" | undefined;
    title?: string | undefined;
    is_active?: boolean | undefined;
    category?: string | undefined;
    description_en?: string | undefined;
    provider?: string | undefined;
    language?: string | undefined;
    is_mandatory?: boolean | undefined;
    tags?: string[] | undefined;
    duration_hours?: number | undefined;
    skill_level?: "beginner" | "intermediate" | "advanced" | "expert" | undefined;
    title_en?: string | undefined;
    is_certification?: boolean | undefined;
    external_url?: string | undefined;
    thumbnail_url?: string | undefined;
    prerequisites?: string[] | undefined;
}, {
    description?: string | undefined;
    status?: "draft" | "published" | "archived" | undefined;
    title?: string | undefined;
    is_active?: boolean | undefined;
    category?: string | undefined;
    description_en?: string | undefined;
    provider?: string | undefined;
    language?: string | undefined;
    is_mandatory?: boolean | undefined;
    tags?: string[] | undefined;
    duration_hours?: number | undefined;
    skill_level?: "beginner" | "intermediate" | "advanced" | "expert" | undefined;
    title_en?: string | undefined;
    is_certification?: boolean | undefined;
    external_url?: string | undefined;
    thumbnail_url?: string | undefined;
    prerequisites?: string[] | undefined;
}>;